Split common TypeConversions utils
