+++ BUILD.bazel
@@ -519,6 +519,8 @@ cc_library(
         "stablehlo/conversions/linalg/transforms/StablehloToLinalgRandom.cpp",
         "stablehlo/conversions/linalg/transforms/StablehloToLinalgReduce.cpp",
         "stablehlo/conversions/linalg/transforms/TypeConversion.cpp",
+        "stablehlo/transforms/conversions/TypeConversion.h",
+        "stablehlo/transforms/conversions/TypeConversion.cpp",
     ],
     hdrs = [
         "stablehlo/conversions/linalg/transforms/LegalizeToLinalgUtils.h",
@@ -1162,7 +1164,6 @@ cc_library(
         ":broadcast_utils",
         ":chlo_ops",
         ":chlo_rewriters_inc_gen",
-        ":linalg_passes",
         ":stablehlo_aggressive_simplification_inc_gen",
         ":stablehlo_create_compatibility_expander_inc_gen",
         ":stablehlo_create_complex_math_expander_inc_gen",
@@ -1171,6 +1172,7 @@ cc_library(
         ":stablehlo_ops_inc_gen",
         ":stablehlo_pass_inc_gen",
         ":stablehlo_type_inference",
+        ":stablehlo_type_conversions",
         ":version",
         ":vhlo_ops",
         ":vhlo_rewriters_inc_gen",
@@ -1242,6 +1244,23 @@ cc_library(
         "@llvm-project//mlir:Parser",
         "@llvm-project//mlir:Pass",
         "@llvm-project//mlir:Support",
+    ],
+)
+
+cc_library(
+    name = "stablehlo_type_conversions",
+    srcs = [
+        "stablehlo/transforms/conversions/TypeConversion.cpp",
+    ],
+    hdrs = [
+        "stablehlo/transforms/conversions/TypeConversion.h",
+    ],
+    strip_include_prefix = ".",
+    deps = [
+        "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:Support",
+        "@llvm-project//mlir:TransformUtils",
     ],
 )
