+++ stablehlo/conversions/linalg/transforms/TypeConversion.cpp
@@ -31,45 +31,11 @@ limitations under the License.
 #include "mlir/IR/Value.h"
 #include "mlir/IR/ValueRange.h"
 #include "mlir/Support/LLVM.h"
+#include "stablehlo/transforms/conversions/TypeConversion.h"
 
 namespace mlir::stablehlo {
 
 namespace {
-
-Type convertInteger(IntegerType intType) {
-  return IntegerType::get(intType.getContext(),
-                          intType.getIntOrFloatBitWidth());
-}
-
-Type convertShapedType(ShapedType shapedType) {
-  if (auto intType = llvm::dyn_cast<IntegerType>(shapedType.getElementType()))
-    return shapedType.clone(convertInteger(intType));
-  return shapedType;
-}
-
-Value materializeCastFromIllegal(OpBuilder &builder, Type type,
-                                 ValueRange inputs, Location loc) {
-  Type fromType = getElementTypeOrSelf(inputs[0].getType());
-  Type toType = getElementTypeOrSelf(type);
-  if ((!fromType.isSignedInteger() && !fromType.isUnsignedInteger()) ||
-      !toType.isSignlessInteger())
-    return Value();
-  // Use unrealized conversion casts to do signful->signless conversions.
-  return builder.create<UnrealizedConversionCastOp>(loc, type, inputs[0])
-      ->getResult(0);
-}
-
-Value materializeCastToIllegal(OpBuilder &builder, Type type, ValueRange inputs,
-                               Location loc) {
-  Type fromType = getElementTypeOrSelf(inputs[0].getType());
-  Type toType = getElementTypeOrSelf(type);
-  if (!fromType.isSignlessInteger() ||
-      (!toType.isSignedInteger() && !toType.isUnsignedInteger()))
-    return Value();
-  // Use unrealized conversion casts to do signless->signful conversions.
-  return builder.create<UnrealizedConversionCastOp>(loc, type, inputs[0])
-      ->getResult(0);
-}
 
 Value scalarToTensor(OpBuilder &builder, Type type, ValueRange inputs,
                      Location loc) {
@@ -93,17 +59,6 @@ Value scalarToTensor(OpBuilder &builder, Type type, ValueRange inputs,
 }
 
 }  // namespace
-
-RemoveSignTypeConverter::RemoveSignTypeConverter() {
-  addConversion([](Type type) { return type; });
-
-  addConversion(convertInteger);
-  addConversion(convertShapedType);
-
-  addArgumentMaterialization(materializeCastFromIllegal);
-  addSourceMaterialization(materializeCastToIllegal);
-  addTargetMaterialization(materializeCastFromIllegal);
-}
 
 LinalgTypeConverter::LinalgTypeConverter() : RemoveSignTypeConverter() {
   addArgumentMaterialization(scalarToTensor);
